CHORE: upgrade addon to ember v 3.12 (#2)
* ENH: add yarn.lock file

Having a .lock file is generally a good idea :)
Also, without it I have to manually set "resolutions" for ember-forecast-io/ember-data in the consuming app.

* ENH: upgrade to ember 3.12

* ENH: run ember-qunit-codemod

To update the test API syntax

* ENH: add lint and js configuration

* ENH: convert addon to native classes

* FIX: async model relationships

After upgrading ember-forecast-io we started receiving errors related to requests made for fetching `data_points`, which in fact are supposed to be embedded. According to EmbeddedRecordsMixin documentation `If a model is declared embedded for deserialization (embedded: 'always' or deserialize: 'records'), then do not use async: true`.  And { async: true } is the default option for relationships.